Added Seperate Counter/s Object Files and Fixed the Deletion Issue
